RavenDB-16865 Add cloud link expire date and rename components
