+++ src/Raven.Studio/typescript/components/pages/resources/databases/partials/create/formBackup/createDatabaseFromBackupValidation.ts
@@ -74,6 +74,7 @@ const ravenCloudSource = yup.object({
             getBackupConfigurationScript: yup.string(),
             customServerUrl: yup.string(),
             forcePathStyle: yup.boolean(),
+            expireDate: yup.string(),
         })
         .nullable(),
 });
